parser/requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.11.8
aiohttp_socks==0.9.1
aiosignal==1.3.1
asyncpg==0.30.0
attrs==24.2.0
et_xmlfile==2.0.0
frozenlist==1.5.0
idna==3.10
multidict==6.1.0
openpyxl==3.1.5
propcache==0.2.0
python-dotenv==1.0.1
python-socks==2.5.3
yarl==1.18.0
